Improvment
Tambah validasi
Tambah Kontak Admin
Tambah pesan jika pedagang belum memiliki toko
front end minor update
etc

+++ app/Controllers/Admin.php
@@ -7,6 +7,7 @@ use CodeIgniter\Shield\Entities\User;
 use CodeIgniter\Shield\Models\UserModel;
 use CodeIgniter\Files\File;
 use App\Models\TokoModel;
+use App\Models\InfoModel;
 
 class Admin extends BaseController
 {
@@ -59,34 +60,99 @@ class Admin extends BaseController
             return redirect()->back()->with('error', 'You do not have permissions to access that page.');
         }
 
+
+
         $users = auth()->getProvider();
         $username = $this->request->getVar('username');
         $email = $this->request->getVar('email');
         $password = password_hash($this->request->getVar('password'), PASSWORD_BCRYPT); // **Hash the password before saving!**
+        $nohp = $this->request->getVar('nohp');
 
         $user = new User([
             'username' => $username,
             'email' => $email,
+            'nohp' => $nohp,
             'password' => $password,
             'user_image' => 'default.svg',
             'active' => 1
         ]);
+
+        // Validation rules
+        $rules = [
+            'username' => 'required|min_length[3]|max_length[20]|is_unique[users.username]',
+            'email' => 'required|min_length[6]|max_length[50]|valid_email|is_unique[auth_identities.secret]',
+            'nohp' => 'required|min_length[10]|max_length[13]',
+            'password' => 'required|min_length[8]|max_length[255]',
+            'password_confirm' => 'matches[password]'
+        ];
+        $messages = [
+            'username' => [
+                'is_unique' => 'Username sudah digunakan'
+            ],
+            'email' => [
+                'is_unique' => 'Email sudah digunakan'
+            ],
+            'password' => [
+                'required' => 'Password tidak boleh kosong'
+            ],
+            'nohp' => [
+                'required' => 'No HP tidak boleh kosong'
+            ],
+            'password_confirm' => [
+                'matches' => 'Password tidak sama'
+            ]
+        ];
+
+        // Run validation
+        if (!$this->validate($rules, $messages)) {
+            // Redirect back with error message
+            if ($this->validator->hasError('username')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('username'));
+            } elseif ($this->validator->hasError('email')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('email'));
+            } elseif ($this->validator->hasError('password')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('password'));
+            } elseif ($this->validator->hasError('password_confirm')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('password_confirm'));
+            } elseif ($this->validator->hasError('nohp')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('nohp'));
+            }
+        }
+
         $users->save($user);
 
         // To get the complete user object with ID, we need to get from the database
         $user = $users->findById($users->getInsertID());
 
-        // Check if user is created successfully
-        if (!$user) {
-            return redirect()->to('/admin')->with('error', 'Failed to create user');
-        }
-
         // Save auth_groups_users
         $group_id = $this->request->getVar('role');
+        //group_id validation
+        $rulesgroup = [
+            'role' => 'required'
+        ];
+
+        $messagesgroup = [
+            'role' => [
+                'required' => 'Role tidak boleh kosong'
+            ]
+        ];
+
+        if (!$this->validate($rulesgroup, $messagesgroup)) {
+            // Redirect back with error message
+            if ($this->validator->hasError('role')) {
+                return redirect()->back()->withInput()->with('error', $this->validator->getError('role'));
+            }
+        }
+
         if ($group_id !== null) {
             $user->addGroup($group_id);
         } else {
             $user->addGroup('user');
+        }
+
+        // Check if user is created successfully
+        if (!$user) {
+            return redirect()->to('/admin')->with('error', 'Failed to create user');
         }
 
 
@@ -149,14 +215,18 @@ class Admin extends BaseController
         if (!auth()->user()->can('users.edit')) {
             return redirect()->back()->with('error', 'You do not have permissions to access that page.');
         }
+
         $id = $this->request->getVar('id');
         $username = $this->request->getVar('username');
         $fullname = $this->request->getVar('fullname');
         $email = $this->request->getVar('email');
         $role = $this->request->getVar('role');
+        $nohp = $this->request->getVar('nohp');
         $users = auth()->getProvider();
 
         $user = $users->findById($id);
+
+
 
         // Handle the file upload
         $file = $this->request->getFile('user_image');
@@ -186,13 +256,26 @@ class Admin extends BaseController
         $user->fill([
             'username' => $username,
             'fullname' => $fullname,
+            'nohp' => $nohp,
             'email' => $email,
             'user_image' => $user_image,
             'password' => $password,
             'group' => $role
         ]);
 
+        $groupbefore = $user->getGroups();
+        // Update user group
+        if (is_array($groupbefore)) {
+            foreach ($groupbefore as $group) {
+                $user->removeGroup($group);
+            }
+        } else {
+            $user->removeGroup($groupbefore);
+        }
+        $user->addGroup($role);
         $users->save($user);
+
+
         session()->setFlashdata('success', 'Data Berhasil Diubah');
         return redirect()->to('/admin');
     }
@@ -204,6 +287,10 @@ class Admin extends BaseController
         }
         $users = new UserModel();
         $user = $users->find($id);
+        $loggedInUserId = Auth()->user()->id;
+        if ($user->id === $loggedInUserId) {
+            return redirect()->to('/admin')->with('error', 'You cannot delete yourself');
+        }
 
         if ($user) {
             $users->delete($id);
@@ -249,7 +336,7 @@ class Admin extends BaseController
 
     public function tambah_toko()
     {
-        
+
         if (!auth()->user()->can('users.create')) {
             return redirect()->back()->with('error', 'You do not have permissions to access that page.');
         }
@@ -265,14 +352,14 @@ class Admin extends BaseController
 
         // Run validation
         if (!$this->validate($rules)) {
-            return redirect()->back()->withInput()->with('error', 'Salah bang');
+            return redirect()->back()->withInput()->with('error', 'Field Tidak Boleh Kosong');
         }
 
         // Get the input data from the form
         $namaToko = $this->request->getVar('nama_toko');
         $alamatToko = $this->request->getVar('alamat_toko');
         $teleponToko = $this->request->getVar('telepon_toko');
-        $idUser = $this->request->getVar('owner'); 
+        $idUser = $this->request->getVar('owner');
 
         // Create a new toko record
         $tokoData = [
@@ -283,7 +370,7 @@ class Admin extends BaseController
         ];
         $tokoModel->insert($tokoData);
         // d($tokoData);
-        
+
 
         session()->setFlashdata('success', 'Toko berhasil ditambahkan');
         return redirect()->to('/admin/toko');
@@ -296,7 +383,7 @@ class Admin extends BaseController
         }
         $tokoModel = new TokoModel();
         $toko = $tokoModel->find($id);
-        
+
         $usersm = new UserModel();
         $users = $usersm->findAll();
 
@@ -313,7 +400,8 @@ class Admin extends BaseController
         }
     }
 
-    public function perbarui_toko(){
+    public function perbarui_toko()
+    {
         if (!auth()->user()->can('users.edit')) {
             return redirect()->back()->with('error', 'You do not have permissions to access that page.');
         }
@@ -399,4 +487,5 @@ class Admin extends BaseController
     }
 
 
+}
 }
